# docs/api/OPENAPI.yaml
openapi: 3.1.0
info:
  title: Caladdin API
  version: 1.0.0
  description: |
    HTTP API for Caladdin — AI scheduling assistant.

    **Route groups covered:** `/api/*`, `/book/*`, `/s/*`, `/jobs/*`

    **Authentication:**
    - **Session cookie** (`caladdin_session`) — host-facing routes under `/api/*`
    - **API key** (`x-api-key` header) — cron/worker routes under `/jobs/*`
    - **Guest action token** (signed HMAC, query or body) — cancel/reschedule on `/s/:token/*`
    - **Public** — `/book/*` metadata and `/s/:token` HTML pages (no auth)

    Base URL is `CALADDIN_BASE_URL` (e.g. `https://caladdin.onrender.com`).
servers:
  - url: https://caladdin.onrender.com
    description: Production (example)
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: caladdin_session
      description: HMAC-signed session cookie set by `GET /auth/callback`
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Value of `CALADDIN_API_KEY` env var
  schemas:
...
    EventType:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        durationMinutes:
          type: integer
          minimum: 1
          maximum: 480
        description:
          type: string
          nullable: true
        availabilityRules:
          type: object
          additionalProperties: true
          description: Weekly schedule, buffers, minimum notice (see PublicBookingEventType)
        schedulingMode:
          type: string
          enum: [single, round_robin]
        active:
          type: boolean
        publicUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EventTypeCreate:
      type: object
      required: [name, durationMinutes]
      properties:
        name:
          type: string
        slug:
          type: string
          description: Optional; auto-generated from name if omitted
        durationMinutes:
          type: integer
          minimum: 1
          maximum: 480
        description:
          type: string
          nullable: true
        availabilityRules:
          type: object
          additionalProperties: true
    EventTypePatch:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        durationMinutes:
          type: integer
          minimum: 1
          maximum: 480
        description:
          type: string
          nullable: true
        availabilityRules:
          type: object
          additionalProperties: true
        active:
          type: boolean
...
    JobComplete:
      type: object
      properties:
        status:
          type: string
          enum: [complete]
paths:
...
  /api/event-types:
    get:
      tags: [Event Types]
      summary: List event types for host
      security:
        - sessionCookie: []
      parameters:
        - name: includeInactive
          in: query
          schema:
            type: boolean
            default: false
          description: Include deactivated event types when `true`
      responses:
        '200':
          description: Event type list
          content:
            application/json:
              schema:
                type: object
                properties:
                  eventTypes:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventType'
        '401':
          description: Not authenticated
    post:
      tags: [Event Types]
      summary: Create event type
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventTypeCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  eventType:
                    $ref: '#/components/schemas/EventType'
        '400':
          description: Validation error
        '401':
          description: Not authenticated
        '409':
          description: Slug already exists
  /api/event-types/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Event Types]
      summary: Get event type by ID
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Event type
          content:
            application/json:
              schema:
                type: object
                properties:
                  eventType:
                    $ref: '#/components/schemas/EventType'
        '401':
          description: Not authenticated
        '404':
          description: Not found
    patch:
      tags: [Event Types]
      summary: Update event type
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventTypePatch'
      responses:
        '200':
          description: Updated event type
          content:
            application/json:
              schema:
                type: object
                properties:
                  eventType:
                    $ref: '#/components/schemas/EventType'
        '400':
          description: Validation error
        '401':
          description: Not authenticated
        '404':
          description: Not found
        '409':
          description: Slug conflict
    delete:
      tags: [Event Types]
      summary: Deactivate event type (soft delete)
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Deactivated event type
          content:
            application/json:
              schema:
                type: object
                properties:
                  eventType:
                    $ref: '#/components/schemas/EventType'
        '401':
          description: Not authenticated
        '404':
          description: Not found
  /api/event-types/{id}/members:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Event Types]
      summary: List round-robin team members for an event type
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Team member user IDs in rotation order
        '401':
          description: Not authenticated
        '404':
          description: Event type not found
    put:
      tags: [Event Types]
      summary: Replace round-robin team members
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                memberUserIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Updated members
        '401':
          description: Not authenticated
        '404':
          description: Event type not found
...
  /jobs/session-expiry:
    post:
      tags: [Jobs]
      summary: Expire open scheduling sessions
      description: |
        Render cron `caladdin-session-expiry` calls this every 15 minutes.
        Also runs in-process on the web service as a fallback worker.
      security:
        - apiKey: []
      responses:
        '200':
          description: Job complete
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/JobComplete'
                  - type: object
                    properties:
                      expired:
                        type: integer
                        description: Number of sessions marked expired
        '401':
          description: Missing or invalid API key
        '500':
          description: Job failed
  /jobs/reminders:
    post:
      tags: [Jobs]
      summary: Send due booking reminder emails
      description: Render cron `caladdin-reminders` calls this hourly.
      security:
        - apiKey: []
      responses:
        '200':
          description: Job complete
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/JobComplete'
                  - type: object
                    properties:
                      processed:
                        type: integer
                      sent:
                        type: integer
                      skipped:
                        type: integer
                      failed:
                        type: integer
        '401':
          description: Missing or invalid API key
        '500':
          description: Job failed
  /jobs/improvement-loop:
    post:
      tags: [Jobs]
      summary: Analyze failures and emit improvement report
      description: Optional manual/scheduled job; not in Render cron by default.
      security:
        - apiKey: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                lookbackDays:
                  type: integer
                  default: 7
                minFailuresPerGroup:
                  type: integer
                  default: 3
      responses:
        '200':
          description: Job complete
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/JobComplete'
                  - type: object
                    properties:
                      failuresAnalyzed:
                        type: integer
                      groupsAnalyzed:
                        type: integer
                      reportPath:
                        type: string
                      ntfySent:
                        type: boolean
        '401':
          description: Missing or invalid API key
        '500':
          description: Job failed